feat: add redirect to chat and refactor folder
