+++ src/components/friends/MyFriendCard.tsx
@@ -1,11 +1,12 @@
 import { Card, Flex, Box, Text, Button, Icon } from "@chakra-ui/react";
-
+import Link from "next/link";
 import { MdOutlineChatBubbleOutline } from "react-icons/md";
 
 export default function MyFriendCard(props: {
   name: string;
   image?: string;
   bio: string;
+  id: string;
 }) {
   return (
     <Card
@@ -109,38 +110,40 @@ export default function MyFriendCard(props: {
             </Text>
         </Flex>
         </Flex>
-        <Button
-          display="flex"
-          flexDirection="row"
-          justifyContent="center"
-          alignItems="center"
-          padding="4px 16px"
-          width="86px"
-          height="23px"
-          gap="12px"
-          bg="#FFFC83"
-          borderRadius="12px"
-        >
-            <Icon
-              width="12px"
-              height="12px"
-              color="#4909B3"
-              as={MdOutlineChatBubbleOutline}
-            >
-            </Icon>
-            {/* Label */}
-            <Text
-              width="30px"
-              height="15px"
-              size="A"
-              fontStyle="normal"
-              fontWeight={1000}
-              lineHeight="15px"
-              color="#4909B3"
-            >
-            CHAT
-            </Text>
-        </Button>
+        <Link href={`chat/${props.id}`}>
+          <Button
+            display="flex"
+            flexDirection="row"
+            justifyContent="center"
+            alignItems="center"
+            padding="4px 16px"
+            width="86px"
+            height="23px"
+            gap="12px"
+            bg="#FFFC83"
+            borderRadius="12px"
+          >
+              <Icon
+                width="12px"
+                height="12px"
+                color="#4909B3"
+                as={MdOutlineChatBubbleOutline}
+              >
+              </Icon>
+              {/* Label */}
+              <Text
+                width="30px"
+                height="15px"
+                size="A"
+                fontStyle="normal"
+                fontWeight={1000}
+                lineHeight="15px"
+                color="#4909B3"
+              >
+              CHAT
+              </Text>
+          </Button>
+        </Link>
     </Flex>
   </Card>
