+++ src/components/friends/SearchedFriends.tsx
