+++ src/constant.tsx
@@ -1,1 +1,1 @@
-export const FUTUREFLAG = false
+export const FUTUREFLAG = true
